Cambios solicitados para el PR
